requirements.txt
streamlit
ultralytics
diffusers
transformers
accelerate
safetensors
sentence-transformers
opencv-python-headless
Pillow
torch
numpy
